+++ homematicip/aio/class_maps.py
@@ -67,6 +67,9 @@ TYPE_CLASS_MAP = {
     DeviceType.WEATHER_SENSOR: AsyncWeatherSensor,
     DeviceType.WEATHER_SENSOR_PLUS: AsyncWeatherSensorPlus,
     DeviceType.WEATHER_SENSOR_PRO: AsyncWeatherSensorPro,
+    DeviceType.WIRED_DIMMER_3: AsyncWiredDimmer3,
+    DeviceType.WIRED_INPUT_32: AsyncWiredInput32,
+    DeviceType.WIRED_SWITCH_8: AsyncWiredSwitch8,
 }
 
 TYPE_GROUP_MAP = {