+++ homematicip/class_maps.py
@@ -70,6 +70,9 @@ TYPE_CLASS_MAP = {
     DeviceType.WEATHER_SENSOR: WeatherSensor,
     DeviceType.WEATHER_SENSOR_PLUS: WeatherSensorPlus,
     DeviceType.WEATHER_SENSOR_PRO: WeatherSensorPro,
+    DeviceType.WIRED_DIMMER_3: WiredDimmer3,
+    DeviceType.WIRED_INPUT_32: WiredInput32,
+    DeviceType.WIRED_SWITCH_8: WiredSwitch8,
 }
 
 TYPE_GROUP_MAP = {