+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### ADDED
 - API
     - Devices
+      - HMIPW-DRD3 (Homematic IP Wired Dimming Actuator – 3x channels)
+      - HMIPW-DRS8 (Homematic IP Wired Switch Actuator – 8x channels)
+      - HMIPW-DRI32 (Homematic IP Wired Inbound module – 32x channels)
       - Fields
         - connectionType
         - new OptionalFeatures